fix(audit): split ci replay owner module
Keep the cold-cache CI replay fix under the install LOC budget, preserve the shared-owner boundary, and align the legacy unit contracts with the new self-hydrating replay path after merging main.


Copilot-Session: f1d182b1-b222-463e-8beb-e3f1334c61e1
