+++ .github/instructions/architecture.instructions.md
@@ -50,7 +50,7 @@ semicolon-delimited, and specific to the file(s) that own the fact.
 | Git repository cache-key normalization | cache/url_normalize.py (normalize_repo_url / cache_shard_key) | `src/apm_cli/cache/url_normalize.py` |
 | Cached policy shape | policy/discovery.py (_policy_to_dict via _serialize_policy) | `src/apm_cli/policy/discovery.py` |
 | Post-uninstall dependency reachability | deps/reachability.py (compute_forward_reachable_keys) | `src/apm_cli/deps/reachability.py` |
-| CI audit scratch materialization | install/drift.py (prepare_ci_audit_replay / run_replay) | `src/apm_cli/install/drift.py` |
+| CI audit scratch materialization | install/audit_replay.py (prepare_ci_audit_replay) | `src/apm_cli/install/audit_replay.py` |
 | GitHub API throttle classification | deps/github_rate_limit.py | `src/apm_cli/deps/github_rate_limit.py` |
 <!-- /canonical-owner-table -->
