+++ scripts/lint-architecture-boundaries.sh
@@ -361,11 +361,11 @@ config_consistency_body=$(awk '
     flag && /^def / && !/^def _check_config_consistency\(/ {exit}
     flag {print}
 ' src/apm_cli/policy/ci_checks.py)
-if ! grep -q '^def prepare_ci_audit_replay(' src/apm_cli/install/drift.py \
+if ! grep -q '^def prepare_ci_audit_replay(' src/apm_cli/install/audit_replay.py \
     || ! printf '%s\n' "$audit_ci_gate_body" | grep -q 'prepare_ci_audit_replay' \
     || printf '%s\n' "$audit_ci_gate_body" | grep -q 'run_replay(' \
     || ! printf '%s\n' "$config_consistency_body" | grep -q 'prepared_replay\.modules_root'; then
-    echo "[x] CI audit scratch materialization must route through install/drift.py"
+    echo "[x] CI audit scratch materialization must route through install/audit_replay.py"
     violations=$((violations + 1))
 fi
 local_bundle_marker_hits=$(
